+++ src/filters/index.js
@@ -1,4 +1,0 @@
-'use strict';
-
-var whiteList = require('./whiteList');
-module.exports = {whiteList: whiteList};