+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 1.6.0 (November 18, 2016)
-- Refactoring enable modules (boost)
+- Refactoring zerg.enable modules (boost)
 
 ## 1.5.0 (November 18, 2016)
 - Return to ES5 for support old browsers